--- Planarian.Web/src/Modules/Account/Models/Archive/BackupDownloadVm.ts
@@ -1,4 +1,0 @@
-export interface BackupDownloadVm {
-    fileName: string;
-    downloadUrl: string;
-}
