--- Planarian.Web/src/Modules/Account/Models/Archive/BackupProgressVm.ts
@@ -1,5 +1,0 @@
-export interface BackupProgressVm {
-    statusMessage: string;
-    processedCaves?: number;
-    totalCaves?: number;
-}
