+++ Planarian.Web/src/Shared/Models/ProgressVm.ts
@@ -1,0 +1,5 @@
+export interface ProgressVm {
+  statusMessage: string;
+  processedCaves?: number;
+  totalCaves?: number;
+}
