+++ Planarian/Planarian/Modules/Account/Archive/Models/ArchiveCaveCsvModel.cs
@@ -1,8 +1,8 @@
 using Planarian.Modules.Import.Models;
 
-namespace Planarian.Modules.Account.Backup.Models;
+namespace Planarian.Modules.Account.Archive.Models;
 
-public sealed class AccountBackupCaveDto : CaveCsvModel
+public sealed class ArchiveCaveCsvModel : CaveCsvModel
 {
     public string PlanarianId { get; set; } = null!;
     public string? CountyIdDelimiter { get; set; }
