+++ Planarian/Planarian/Modules/Account/Archive/Models/ArchiveEntranceByCaveCsvModel.cs
@@ -1,0 +1,6 @@
+namespace Planarian.Modules.Account.Archive.Models;
+
+public sealed class ArchiveEntranceByCaveCsvModel : ArchiveEntranceCsvModel
+{
+    public string CavePlanarianId { get; set; } = null!;
+}
