+++ Planarian/Planarian/Modules/Account/Archive/Models/ArchiveEntranceCsvModel.cs
@@ -1,0 +1,7 @@
+using Planarian.Modules.Import.Models;
+
+namespace Planarian.Modules.Account.Archive.Models;
+
+public class ArchiveEntranceCsvModel : EntranceCsvModel
+{
+}
