+++ Planarian/Planarian/Modules/Account/Archive/Models/ArchiveFileByCaveModel.cs
@@ -1,0 +1,6 @@
+namespace Planarian.Modules.Account.Archive.Models;
+
+public sealed class ArchiveFileByCaveModel : ArchiveFileModel
+{
+    public string CavePlanarianId { get; set; } = null!;
+}
