+++ Planarian/Planarian/Modules/Account/Archive/Models/ArchiveGeoJsonByCaveModel.cs
@@ -1,0 +1,6 @@
+namespace Planarian.Modules.Account.Archive.Models;
+
+public sealed class ArchiveGeoJsonByCaveModel : ArchiveGeoJsonModel
+{
+    public string CavePlanarianId { get; set; } = null!;
+}
