+++ Planarian/Planarian/Modules/Account/Archive/Models/ArchiveGeoJsonModel.cs
@@ -1,0 +1,8 @@
+namespace Planarian.Modules.Account.Archive.Models;
+
+public class ArchiveGeoJsonModel
+{
+    public string Id { get; set; } = null!;
+    public string Name { get; set; } = null!;
+    public string GeoJson { get; set; } = string.Empty;
+}
