--- Planarian/Planarian/Modules/Account/Backup/Models/AccountBackupDownloadVm.cs
@@ -1,7 +1,0 @@
-namespace Planarian.Modules.Account.Backup.Models;
-
-public sealed class AccountBackupDownloadVm
-{
-    public string FileName { get; set; } = null!;
-    public string DownloadUrl { get; set; } = null!;
-}
