--- Planarian/Planarian/Modules/Account/Backup/Models/AccountBackupEntranceByCaveDto.cs
@@ -1,6 +1,0 @@
-namespace Planarian.Modules.Account.Backup.Models;
-
-public sealed class AccountBackupEntranceByCaveDto : AccountBackupEntranceDto
-{
-    public string CavePlanarianId { get; set; } = null!;
-}
