--- Planarian/Planarian/Modules/Account/Backup/Models/AccountBackupEntranceDto.cs
@@ -1,7 +1,0 @@
-using Planarian.Modules.Import.Models;
-
-namespace Planarian.Modules.Account.Backup.Models;
-
-public class AccountBackupEntranceDto : EntranceCsvModel
-{
-}
