--- Planarian/Planarian/Modules/Account/Backup/Models/AccountBackupFileByCaveDto.cs
@@ -1,6 +1,0 @@
-namespace Planarian.Modules.Account.Backup.Models;
-
-public sealed class AccountBackupFileByCaveDto : AccountBackupFileDto
-{
-    public string CavePlanarianId { get; set; } = null!;
-}
