--- Planarian/Planarian/Modules/Account/Backup/Models/AccountBackupGeoJsonByCaveDto.cs
@@ -1,6 +1,0 @@
-namespace Planarian.Modules.Account.Backup.Models;
-
-public sealed class AccountBackupGeoJsonByCaveDto : AccountBackupGeoJsonDto
-{
-    public string CavePlanarianId { get; set; } = null!;
-}
