--- Planarian/Planarian/Modules/Account/Backup/Models/AccountBackupGeoJsonDto.cs
@@ -1,7 +1,0 @@
-namespace Planarian.Modules.Account.Backup.Models;
-
-public class AccountBackupGeoJsonDto
-{
-    public string Name { get; set; } = null!;
-    public string GeoJson { get; set; } = string.Empty;
-}
