+++ Planarian/Planarian/Modules/Account/Controller/AccountController.cs
@@ -1,11 +1,8 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Caching.Memory;
 using Planarian.Library.Extensions.String;
-using Planarian.Library.Exceptions;
 using Planarian.Model.Database.Entities.RidgeWalker;
 using Planarian.Model.Shared;
-using Planarian.Modules.Account.Backup.Models;
 using Planarian.Modules.Account.Model;
 using Planarian.Modules.Account.Services;
 using Planarian.Modules.Authentication.Services;
@@ -19,18 +16,14 @@ namespace Planarian.Modules.Account.Controller;
 public class AccountController : PlanarianControllerBase<AccountService>
 {
     private const string Route = "api/account";
-    private const int MaxBackupRequestsPerDay = 10;
-    private static readonly TimeSpan BackupRequestWindow = TimeSpan.FromDays(1);
 
     private readonly ImportService _importService;
-    private readonly MemoryCache _cache;
 
     public AccountController(RequestUser requestUser, TokenService tokenService, AccountService service,
-        CaveService caveService, ImportService importService, MemoryCache cache) : base(
+        CaveService caveService, ImportService importService) : base(
         requestUser, tokenService, service)
     {
         _importService = importService;
-        _cache = cache;
     }
 
     [HttpPost]
@@ -52,40 +45,15 @@ public class AccountController : PlanarianControllerBase<AccountService>
         return Ok("Account reset finished.");
     }
 
-    #region Backup
+    #region Archive
 
-    [HttpGet("backup")]
+    [HttpGet("archive")]
     [Authorize(Policy = PermissionPolicyKey.Admin)]
-    public async Task<ActionResult<AccountBackupDownloadVm>> DownloadBackup(string? uuid, CancellationToken cancellationToken)
+    public async Task<IActionResult> DownloadArchive(string? uuid, CancellationToken cancellationToken)
     {
-        var cacheKey = GetBackupRateLimitKey();
-        var rateLimitState = _cache.GetOrCreate(cacheKey, entry =>
-        {
-            entry.AbsoluteExpirationRelativeToNow = BackupRequestWindow;
-            return new BackupRateLimitState();
-        })!;
-
-        lock (rateLimitState.SyncRoot)
-        {
-            if (rateLimitState.Attempts >= MaxBackupRequestsPerDay)
-                throw ApiExceptionDictionary.TooManyRequests("Too many backup requests. Try again tomorrow.");
-
-            rateLimitState.Attempts++;
-        }
-
-        var result = await Service.BuildBackup(uuid, cancellationToken);
-        return Ok(result);
-    }
-
-    private string GetBackupRateLimitKey()
-    {
-        return $"backup-rate-limit:{RequestUser.Id}";
-    }
-
-    private sealed class BackupRateLimitState
-    {
-        public int Attempts { get; set; }
-        public object SyncRoot { get; } = new();
+        var fileName = await Service.GetArchiveFileName(cancellationToken);
+        var archiveStream = await Service.BuildArchive(uuid, cancellationToken);
+        return File(archiveStream, "application/zip", fileName);
     }
 
     #endregion
