--- Planarian/Planarian/Shared/Options/BackupOptions.cs
@@ -1,11 +1,0 @@
-namespace Planarian.Shared.Options;
-
-public class BackupOptions
-{
-    public const string Key = "Backup";
-
-    public int BlobDownloadTransferConcurrency { get; set; } = 8;
-    public int BlobDownloadTransferSizeMb { get; set; } = 8;
-    public int TempBlobCleanupIntervalMinutes { get; set; } = 30;
-    public string? TempDirectory { get; set; }
-}
