+++ Planarian/Planarian/Shared/Options/FileOptions.cs
@@ -5,5 +5,4 @@ public class FileOptions
     public const string Key = "File";
 
     public string ConnectionString { get; set; } = null!;
-    public int TempExpirationHours { get; set; } = 2;
 }
