+++ Planarian/Planarian/appsettings.Template.Development.json
@@ -20,13 +20,6 @@
     "ContainerName": "CONTAINER_NAME"
   },
   "File": {
-    "ConnectionString": "FILE_CONNECTION_STRING",
-    "TempExpirationHours": 2
-  },
-  "Backup": {
-    "BlobDownloadTransferConcurrency": 8,
-    "BlobDownloadTransferSizeMb": 8,
-    "TempBlobCleanupIntervalMinutes": 30,
-    "TempDirectory": null
+    "ConnectionString": "FILE_CONNECTION_STRING"
   }
 }
